requirements.txt
# python=3.12
watchdog
chromadb
sentence-transformers
torch
fastapi
uvicorn
pyjwt
pydantic
pydantic-settings
python-multipart